# Only read necessary elements in Chapel array for Parquet